+++ src/DTO/QuotationDTO.php
@@ -81,9 +81,9 @@ final class QuotationDTO extends AbstractDTO
      * @param bool        $recordFreeTextInheritance      True if the free text is inherited.
      * @param string|null $recordOpening                  HTML opening text on the record.
      * @param bool        $recordOpeningInheritance       True if the opening is inherited.
-     * @param AddressDTO|null  $deliveryAddress           Delivery address for this quotation.
-     * @param AddressDTO|null  $invoiceAddress            Invoice address for this quotation.
-     * @param AddressDTO|null  $recordAddress             Record address for this quotation.
+     * @param RecordAddressDTO|null $deliveryAddress       Delivery address for this quotation.
+     * @param RecordAddressDTO|null $invoiceAddress       Invoice address for this quotation.
+     * @param RecordAddressDTO|null $recordAddress        Record address for this quotation.
      * @param string|null                     $dispatchCountryCode     Country code used for dispatch (enum: country).
      * @param list<QuotationItemDTO>          $quotationItems          Line items of this quotation.
      * @param list<CommissionSalesPartnerDTO> $commissionSalesPartners Commission assignments.
@@ -184,9 +184,9 @@ final class QuotationDTO extends AbstractDTO
         public readonly bool        $recordOpeningInheritance,
 
         // Addresses
-        public readonly ?AddressDTO $deliveryAddress,
-        public readonly ?AddressDTO $invoiceAddress,
-        public readonly ?AddressDTO $recordAddress,
+        public readonly ?RecordAddressDTO $deliveryAddress,
+        public readonly ?RecordAddressDTO $invoiceAddress,
+        public readonly ?RecordAddressDTO $recordAddress,
 
         // Dispatch
         public readonly ?string     $dispatchCountryCode,
@@ -215,15 +215,15 @@ final class QuotationDTO extends AbstractDTO
     public static function fromArray(array $data): static
     {
         $deliveryAddress = isset($data['deliveryAddress']) && is_array($data['deliveryAddress'])
-            ? AddressDTO::fromArray($data['deliveryAddress'])
+            ? RecordAddressDTO::fromArray($data['deliveryAddress'])
             : null;
 
         $invoiceAddress = isset($data['invoiceAddress']) && is_array($data['invoiceAddress'])
-            ? AddressDTO::fromArray($data['invoiceAddress'])
+            ? RecordAddressDTO::fromArray($data['invoiceAddress'])
             : null;
 
         $recordAddress = isset($data['recordAddress']) && is_array($data['recordAddress'])
-            ? AddressDTO::fromArray($data['recordAddress'])
+            ? RecordAddressDTO::fromArray($data['recordAddress'])
             : null;
 
         return new static(